release: v0.2.4
All the schema fields gets passed as `required` in the response_format
schema by default. This by itself isn't an issue, but we might want to
specify which of them are required.

Allow to pass an option in `Mentor.Ecto.Schema` to specify which fields
are required `required_fields`

The behaviour is similar to the already existing `ignored_fields`, so to
me made sense to implement this way.