release: v0.2.5
Gemini expects optional fields not to be passed as `required` in the
`response_schema`. Differently from OpenAI which expects all the schema
fields to be passed.

Update the response_schema before sending to Gemini, checking the
`required_fields` from the adapter config.

Since this approach will vary depending on the provider, implemented a
simple private method `update_schema_for_gemini` that updates the schema
when building the request body.